move dashboard to dashboard page and reset home page
